+++ src/ui/src/app/components/gm-live-reload-panel.ts
@@ -4,112 +4,12 @@ import type {
     GraphVisualizationLiveReloadRecentError,
     GraphVisualizationLiveReloadRecentPatch,
     GraphVisualizationLiveReloadRuntimeHealth,
-    GraphVisualizationLiveReloadStatusSnapshot,
-    GraphVisualizationLiveReloadWatcherStatus
+    GraphVisualizationLiveReloadStatusSnapshot
 } from "../../graph/types.js";
 import type { GraphVisualizationUiModel } from "../contracts.js";
-import { getUiErrorMessage } from "../error-message.js";
 import type { GraphVisualizationUiState } from "../state/types.js";
 import { LightDomLitElement } from "./light-dom-lit-element.js";
-
-const DEFAULT_POLL_INTERVAL_MS = 2000;
-const MIN_POLL_INTERVAL_MS = 500;
-
-type UnknownRecord = Readonly<Record<string, unknown>>;
-
-function isUnknownRecord(value: unknown): value is UnknownRecord {
-    return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function readNumber(record: UnknownRecord, key: string): number | null {
-    const value = record[key];
-    return typeof value === "number" && Number.isFinite(value) ? value : null;
-}
-
-function readString(record: UnknownRecord, key: string): string | null {
-    const value = record[key];
-    return typeof value === "string" ? value : null;
-}
-
-function readBoolean(record: UnknownRecord, key: string): boolean | null {
-    const value = record[key];
-    return typeof value === "boolean" ? value : null;
-}
-
-function readRecentPatches(value: unknown): ReadonlyArray<GraphVisualizationLiveReloadRecentPatch> {
-    if (!Array.isArray(value)) {
-        return [];
-    }
-
-    return value.filter(isUnknownRecord).map((entry) => ({
-        durationMs: readNumber(entry, "durationMs") ?? 0,
-        filePath: readString(entry, "filePath") ?? "unknown",
-        hotReloadLatencyMs: readNumber(entry, "hotReloadLatencyMs"),
-        id: readString(entry, "id") ?? "unknown",
-        timestamp: readNumber(entry, "timestamp") ?? 0
-    }));
-}
-
-function readRecentErrors(value: unknown): ReadonlyArray<GraphVisualizationLiveReloadRecentError> {
-    if (!Array.isArray(value)) {
-        return [];
-    }
-
-    return value.filter(isUnknownRecord).map((entry) => ({
-        error: readString(entry, "error") ?? "Unknown error",
-        filePath: readString(entry, "filePath") ?? "unknown",
-        recoveryHint: readString(entry, "recoveryHint"),
-        timestamp: readNumber(entry, "timestamp") ?? 0
-    }));
-}
-
-function resolveWatcherStatus(
-    snapshot: Pick<GraphVisualizationLiveReloadStatusSnapshot, "errorCount" | "scanComplete">,
-    hasStatusUrl: boolean
-): GraphVisualizationLiveReloadWatcherStatus {
-    if (!hasStatusUrl) {
-        return "inactive";
-    }
-
-    if (snapshot.errorCount > 0 && snapshot.scanComplete === false) {
-        return "error";
-    }
-
-    return snapshot.scanComplete ? "running" : "scanning";
-}
-
-function normalizeStatusSnapshot(
-    value: unknown,
-    hasStatusUrl: boolean
-): GraphVisualizationLiveReloadStatusSnapshot | null {
-    if (!isUnknownRecord(value)) {
-        return null;
-    }
-
-    const errorCount = readNumber(value, "errorCount") ?? 0;
-    const scanComplete = readBoolean(value, "scanComplete") ?? false;
-    const baseSnapshot = {
-        errorCount,
-        scanComplete
-    };
-
-    return {
-        avgHotReloadLatencyMs: readNumber(value, "avgHotReloadLatencyMs"),
-        errorCount,
-        maxPatchHistory: readNumber(value, "maxPatchHistory"),
-        patchCount: readNumber(value, "patchCount") ?? 0,
-        patchHistorySize: readNumber(value, "patchHistorySize"),
-        p95HotReloadLatencyMs: readNumber(value, "p95HotReloadLatencyMs"),
-        recentErrors: readRecentErrors(value.recentErrors),
-        recentPatches: readRecentPatches(value.recentPatches),
-        runtimeUrl: readString(value, "runtimeUrl"),
-        scanComplete,
-        totalPatchCount: readNumber(value, "totalPatchCount"),
-        uptimeMs: readNumber(value, "uptime") ?? 0,
-        watcherStatus: resolveWatcherStatus(baseSnapshot, hasStatusUrl),
-        websocketClients: readNumber(value, "websocketClients") ?? 0
-    };
-}
+import { LiveReloadPollingController } from "./live-reload-polling-controller.js";
 
 function formatDurationMs(value: number | null): string {
     if (value === null) {
@@ -152,122 +52,39 @@ export class GmLiveReloadPanel extends LightDomLitElement {
 
     public accessor state: GraphVisualizationUiState | null = null;
 
-    #pollTimer: ReturnType<typeof globalThis.setInterval> | null = null;
-
-    #polledStatus: GraphVisualizationLiveReloadStatusSnapshot | null = null;
-
     #pollErrorMessage: string | null = null;
 
-    #lastStatusUrl: string | null = null;
+    #polledStatus: GraphVisualizationLiveReloadStatusSnapshot | null = null;
 
     #onDismissErrorBanner = (): void => {
         this.#pollErrorMessage = null;
     };
 
+    #pollingController = new LiveReloadPollingController(this, {
+        onErrorMessageChange: (message: string | null): void => {
+            this.#pollErrorMessage = message;
+        },
+        onStatusChange: (status: GraphVisualizationLiveReloadStatusSnapshot | null): void => {
+            this.#polledStatus = status;
+        },
+        requestUpdate: (): void => {
+            this.requestUpdate();
+        }
+    });
+
     public connectedCallback(): void {
         super.connectedCallback();
         this.addEventListener("gm-error-banner-dismiss", this.#onDismissErrorBanner);
-        document.addEventListener("visibilitychange", this.#onVisibilityChange);
-        this.#restartPollingIfNeeded();
     }
 
     public disconnectedCallback(): void {
-        this.#stopPolling();
-        document.removeEventListener("visibilitychange", this.#onVisibilityChange);
         this.removeEventListener("gm-error-banner-dismiss", this.#onDismissErrorBanner);
         super.disconnectedCallback();
     }
 
-    #onVisibilityChange = (): void => {
-        if (document.visibilityState === "visible") {
-            void this.#pollStatusUrlIfNeeded();
-        }
-    };
-
     protected updated(): void {
-        this.#restartPollingIfNeeded();
-    }
-
-    #stopPolling(): void {
-        if (this.#pollTimer !== null) {
-            globalThis.clearInterval(this.#pollTimer);
-            this.#pollTimer = null;
-        }
-    }
-
-    #restartPollingIfNeeded(): void {
         const statusUrl = this.model?.liveReload?.endpoints.statusUrl ?? null;
-        if (statusUrl === this.#lastStatusUrl && this.#pollTimer !== null) {
-            return;
-        }
-
-        this.#stopPolling();
-        this.#lastStatusUrl = statusUrl;
-
-        if (statusUrl === null) {
-            return;
-        }
-
-        void this.#pollStatusUrl(statusUrl);
-        const pollIntervalMs = Math.max(
-            this.model?.liveReload?.pollIntervalMs ?? DEFAULT_POLL_INTERVAL_MS,
-            MIN_POLL_INTERVAL_MS
-        );
-        this.#pollTimer = globalThis.setInterval(() => {
-            void this.#pollStatusUrl(statusUrl);
-        }, pollIntervalMs);
-    }
-
-    async #pollStatusUrl(statusUrl: string): Promise<void> {
-        try {
-            const response = await fetch(statusUrl, {
-                headers: { Accept: "application/json" }
-            });
-            if (!response.ok) {
-                throw new Error(`Status request failed with HTTP ${String(response.status)}`);
-            }
-
-            const payload: unknown = await response.json();
-            const snapshot = normalizeStatusSnapshot(payload, true);
-            if (snapshot === null) {
-                throw new Error("Status response did not match the live-reload snapshot shape.");
-            }
-
-            this.#polledStatus = snapshot;
-            this.#pollErrorMessage = null;
-        } catch (error) {
-            this.#pollErrorMessage = getUiErrorMessage(error, "Failed to refresh live-reload status.");
-        }
-
-        this.requestUpdate();
-    }
-
-    async #pollStatusUrlIfNeeded(): Promise<void> {
-        const statusUrl = this.model?.liveReload?.endpoints.statusUrl ?? null;
-        if (statusUrl === null) {
-            return;
-        }
-
-        try {
-            const response = await fetch(statusUrl, {
-                headers: { Accept: "application/json" }
-            });
-            if (!response.ok) {
-                return;
-            }
-
-            const payload: unknown = await response.json();
-            const snapshot = normalizeStatusSnapshot(payload, true);
-            if (snapshot === null) {
-                return;
-            }
-
-            this.#polledStatus = snapshot;
-            this.#pollErrorMessage = null;
-            this.requestUpdate();
-        } catch {
-            // Silently ignore focus-triggered poll errors.
-        }
+        this.#pollingController.restartPollingIfNeeded(statusUrl, this.model?.liveReload?.pollIntervalMs);
     }
 
     #resolveStatusSnapshot(): GraphVisualizationLiveReloadStatusSnapshot | null {
