+++ src/ui/src/app/components/index.ts
@@ -10,5 +10,7 @@ export * from "./gm-graph-toolbar.js";
 export * from "./gm-live-reload-panel.js";
 export * from "./gm-mcp-panel.js";
 export * from "./gm-playground-panel.js";
+export * from "./lifecycle-participants-controller.js";
 export * from "./light-dom-lit-element.js";
+export * from "./live-reload-polling-controller.js";
 export * from "./primitives/index.js";
